Enemy movement added

+++ Assets/Scripts/Bounds.cs
@@ -4,6 +4,6 @@ using System.Collections.Generic;
 [System.Serializable]
 public class Bounds 
 {
-    public int max;
-    public int min;
+    public float max;
+    public float min;
 }
